requirements.txt
pillow
PyQt5

